🎨 Refactor Coffee Spot Suggestion Functionality
The coffee suggestion feature was the last remaining "ability" of the
sugarcity bot that required removal from the events package. This
commit moves this feature to it's own package.
